simple-adk-agent/pyproject.toml
[project]
name = "simple-adk-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.22.0",
]
